+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Extended Constellation unit tests
    * Standardized Instrument instantiation to always define `inst_module`
    * Extended testing options for `pysat.utils.testing` functions
+   * Added `start_time` keyword for test instruments
 * Deprecations
    * Removed `freq` as a standard kwarg for `pysat.Instruments.download`
 * Documentation
@@ -37,6 +38,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Added missing unit tests for `pysat.utils.time`
    * Added missing unit tests for `pysat.utils.file.parse_delimited_filename`
    * Streamlined unit tests for `test_orbits`
+   * Moved metadata generation for test instruments to `methods.testing`
    * Updated class declaration to be consistent with python 3 standards
 
 [3.0.1] - 2021-07-28